Move TestNetworkService to util package
