+++ code/test-utils/src/main/java/com/adobe/marketing/mobile/util/TestHelper.java
@@ -30,7 +30,6 @@ import com.adobe.marketing.mobile.LoggingMode;
 import com.adobe.marketing.mobile.MobileCore;
 import com.adobe.marketing.mobile.MobileCoreHelper;
 import com.adobe.marketing.mobile.services.MockDataStoreService;
-import com.adobe.marketing.mobile.services.TestNetworkService;
 import com.adobe.marketing.mobile.services.HttpConnecting;
 import com.adobe.marketing.mobile.services.HttpMethod;
 import com.adobe.marketing.mobile.services.Log;
